Implement cuGetProcAddress
This is a function checker that grabs the memory address for the various
cuda functions. Since the "real" cuda functions cuXXX do not exist as own
functions in the library, this has to be "translated" to output the
addresses for the wine_cuXX functions instead, and afaik there is no easy
way to do this without a "ugly list of functions", since you can't really
use wildcards or string manipulation to grab memory addresses of functions
directly.
This is mostly used by CUDA Runtime API calls like cudaGetDeviceCount and
similar that does multiple cuda drivercalls behind the scenes.

Currently the "flags" value is not handled correctly