Move cuGetProcAddress function retriever to separate source
This tidy up the massive amount of "conversion" that needs to be done
to use the wine_xx version of the memory address. Putting it in a
separate sourcefile will make it somewhat easier to add new requests
there in the case of newer SDK expanding this "getProcAddress" funcion
for new added calls.

This getProcAddress call was first added in CUDA 11.3, and then expanded
for CUDA 12. CUDA Runtime API >= 11.6 is still not supported, so verification
there of various cuda samples has not yet been done properly.